Extracted legacy code in own package
